Now using submodels for propulsion mass breakdown
