+++ src/fastoad/models/weight/mass_breakdown/b_propulsion/constants.py
@@ -1,0 +1,17 @@
+"""Constants for propulsion mass submodels."""
+#  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
+#  FAST is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+SERVICE_ENGINE_MASS = "service.mass.propulsion.engine"
+SERVICE_FUEL_LINES_MASS = "service.mass.propulsion.fuel_lines"
+SERVICE_UNCONSUMABLES_MASS = "service.mass.propulsion.unconsumables"
